# src/openapi/packages/star-fleet-commander/Model/SpaceshipRequest.php
<?php
/**
 * SpaceshipRequest
 *
 * PHP version 7.1.3
 *
 * @category Class
 * @package  OpenAPI\Server\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * Star Fleet Commander API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: aleksandar.rakic@yahoo.com
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace OpenAPI\Server\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the SpaceshipRequest model.
 *
 * @package OpenAPI\Server\Model
 * @author  OpenAPI Generator team
 */
class SpaceshipRequest 
{
        /**
     * @var string
     * @SerializedName("spaceship")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $spaceship;

    /**
     * @var bool
     * @SerializedName("captain_junior")
     * @Assert\NotNull()
     * @Assert\Type("bool")
     * @Type("bool")
     */
    protected $captainJunior;

    /**
     * @var int
     * @SerializedName("quantity")
     * @Assert\NotNull()
     * @Assert\Type("int")
     * @Type("int")
     */
    protected $quantity;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->spaceship = isset($data['spaceship']) ? $data['spaceship'] : null;
        $this->captainJunior = isset($data['captainJunior']) ? $data['captainJunior'] : null;
        $this->quantity = isset($data['quantity']) ? $data['quantity'] : null;
    }

    /**
     * Gets spaceship.
     *
     * @return string
     */
    public function getSpaceship()
    {
        return $this->spaceship;
    }

    /**
     * Sets spaceship.
     *
     * @param string $spaceship
     *
     * @return $this
     */
    public function setSpaceship($spaceship)
    {
        $this->spaceship = $spaceship;

        return $this;
    }

    /**
     * Gets captainJunior.
     *
     * @return bool
     */
    public function isCaptainJunior()
    {
        return $this->captainJunior;
    }

    /**
     * Sets captainJunior.
     *
     * @param bool $captainJunior
     *
     * @return $this
     */
    public function setCaptainJunior($captainJunior)
    {
        $this->captainJunior = $captainJunior;

        return $this;
    }

    /**
     * Gets quantity.
     *
     * @return int
     */
    public function getQuantity()
    {
        return $this->quantity;
    }

    /**
     * Sets quantity.
     *
     * @param int $quantity
     *
     * @return $this
     */
    public function setQuantity($quantity)
    {
        $this->quantity = $quantity;

        return $this;
    }
}