# src/openapi/packages/star-fleet-commander/Model/SpaceshipResponse.php
<?php
/**
 * SpaceshipResponse
 *
 * PHP version 7.1.3
 *
 * @category Class
 * @package  OpenAPI\Server\Model
 * @author   OpenAPI Generator team
 * @link     https://github.com/openapitools/openapi-generator
 */

/**
 * Star Fleet Commander API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: aleksandar.rakic@yahoo.com
 * Generated by: https://github.com/openapitools/openapi-generator.git
 *
 */

/**
 * NOTE: This class is auto generated by the openapi generator program.
 * https://github.com/openapitools/openapi-generator
 * Do not edit the class manually.
 */

namespace OpenAPI\Server\Model;

use Symfony\Component\Validator\Constraints as Assert;
use JMS\Serializer\Annotation\Type;
use JMS\Serializer\Annotation\SerializedName;

/**
 * Class representing the SpaceshipResponse model.
 *
 * @package OpenAPI\Server\Model
 * @author  OpenAPI Generator team
 */
class SpaceshipResponse 
{
        /**
     * @var string
     * @SerializedName("uuid")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $uuid;

    /**
     * @var string
     * @SerializedName("name")
     * @Assert\NotNull()
     * @Assert\Type("string")
     * @Type("string")
     */
    protected $name;

    /**
     * @var bool
     * @SerializedName("military")
     * @Assert\NotNull()
     * @Assert\Type("bool")
     * @Type("bool")
     */
    protected $military;

    /**
     * @var bool
     * @SerializedName("captain_junior")
     * @Assert\NotNull()
     * @Assert\Type("bool")
     * @Type("bool")
     */
    protected $captainJunior;

    /**
     * @var int
     * @SerializedName("rank")
     * @Assert\NotNull()
     * @Assert\Type("int")
     * @Type("int")
     */
    protected $rank;

    /**
     * Constructor
     * @param mixed[] $data Associated array of property values initializing the model
     */
    public function __construct(array $data = null)
    {
        $this->uuid = isset($data['uuid']) ? $data['uuid'] : null;
        $this->name = isset($data['name']) ? $data['name'] : null;
        $this->military = isset($data['military']) ? $data['military'] : null;
        $this->captainJunior = isset($data['captainJunior']) ? $data['captainJunior'] : null;
        $this->rank = isset($data['rank']) ? $data['rank'] : null;
    }

    /**
     * Gets uuid.
     *
     * @return string
     */
    public function getUuid()
    {
        return $this->uuid;
    }

    /**
     * Sets uuid.
     *
     * @param string $uuid
     *
     * @return $this
     */
    public function setUuid($uuid)
    {
        $this->uuid = $uuid;

        return $this;
    }

    /**
     * Gets name.
     *
     * @return string
     */
    public function getName()
    {
        return $this->name;
    }

    /**
     * Sets name.
     *
     * @param string $name
     *
     * @return $this
     */
    public function setName($name)
    {
        $this->name = $name;

        return $this;
    }

    /**
     * Gets military.
     *
     * @return bool
     */
    public function isMilitary()
    {
        return $this->military;
    }

    /**
     * Sets military.
     *
     * @param bool $military
     *
     * @return $this
     */
    public function setMilitary($military)
    {
        $this->military = $military;

        return $this;
    }

    /**
     * Gets captainJunior.
     *
     * @return bool
     */
    public function isCaptainJunior()
    {
        return $this->captainJunior;
    }

    /**
     * Sets captainJunior.
     *
     * @param bool $captainJunior
     *
     * @return $this
     */
    public function setCaptainJunior($captainJunior)
    {
        $this->captainJunior = $captainJunior;

        return $this;
    }

    /**
     * Gets rank.
     *
     * @return int
     */
    public function getRank()
    {
        return $this->rank;
    }

    /**
     * Sets rank.
     *
     * @param int $rank
     *
     * @return $this
     */
    public function setRank($rank)
    {
        $this->rank = $rank;

        return $this;
    }
}